Update find-uid-for-rce-vuln.java
If you return true unconditionally at the end of the method, it means that regardless of the mime type and the presence of "uid" in the decoded body, the method will always return true. This implies that the JSON content and the presence of "uid" are not considered for the final decision, and the method will always succeed (return true).